rename: CelebInfoGridCard.jsx 파일 위치 변경 @/components/common => @/components/celebrity
