+++ src/src/components/celebrity/CelebInfoGridCard.jsx
